return token on confirmation
