+++ auth-server/src/resolvers/mutation/addressLinkConfirm.ts
@@ -1,10 +1,5 @@
-import { ForbiddenError } from 'apollo-server';
-import { Keyring } from '@polkadot/api';
-import { schnorrkelVerify } from '@polkadot/util-crypto';
-
-import Address from '../../model/Address';
 import AuthService from '../../services/auth';
-import { Context, MessageType } from '../../types';
+import { Context, ChangeResponseType } from '../../types';
 import getTokenFromReq from '../../utils/getTokenFromReq';
 import messages from '../../utils/messages';
 
@@ -13,41 +8,11 @@ interface argsType {
 	signature: string
 }
 
-const keyring = new Keyring({ type: 'sr25519' });
-
-export default async (parent, { address_id, signature }: argsType, ctx: Context): Promise<MessageType>  => {
-	const token = getTokenFromReq(ctx.req);
+export default async (parent, { address_id, signature }: argsType, ctx: Context): Promise<ChangeResponseType>  => {
+	let token = getTokenFromReq(ctx.req);
 	const authServiceInstance = new AuthService();
-	const user = await authServiceInstance.GetUser(token);
 
-	const dbAddress = await Address
-		.query()
-		.where('id', address_id)
-		.first();
+	token = await authServiceInstance.AddressLinkConfirm(token, address_id, signature);
 
-	if (!dbAddress) {
-		throw new ForbiddenError(messages.ADDRESS_NOT_FOUND);
-	}
-
-	if (dbAddress.user_id !== user.id) {
-		throw new ForbiddenError(messages.ADDRESS_USER_NOT_MATCHING);
-	}
-
-	const publicKey = keyring.decodeAddress(dbAddress.address);
-
-	const isValidSr = schnorrkelVerify(dbAddress.sign_message, signature, publicKey);
-
-	if (!isValidSr) {
-		throw new ForbiddenError(messages.ADDRESS_LINKING_FAILED);
-	}
-
-	await Address
-		.query()
-		.patch({
-			public_key: Buffer.from(publicKey).toString('hex'),
-			verified: true
-		})
-		.findById(address_id);
-
-	return { message: messages.ADDRESS_LINKING_SUCCESSFUL };
+	return { message: messages.ADDRESS_LINKING_SUCCESSFUL, token };
 };
